requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
plotly>=5.17.0
reportlab>=4.0.0
python-dotenv>=1.0.0
bcrypt>=4.1.0

